pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-deploy-tool"
version = "0.1.0"
description = "Stateless CLI for blueprint-driven model deployment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "PyYAML>=6.0.1",
  "requests>=2.31.0",
  "rich>=13.7.1"
]

[project.scripts]
mdp = "mdp_cli.main:run"

[tool.setuptools]
packages = ["mdp_cli"]

[tool.pytest.ini_options]
pythonpath = ["."]
